Fix typo for jinja example and converted to dir
