+++ examples/jinja_example/templates/example_template.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <title>My Webpage</title>
+</head>
+<body>
+    <h1>Hello World</h1>
+    <p>knights - {{ knights }}</p>
+</body>
+</html>
